requirements.txt
yfinance>=1.1.0
pyyaml>=6.0
numpy>=1.24.0
